+++ lib/langchain.rb
@@ -1,8 +1,34 @@
 # frozen_string_literal: true
 
+require "logger"
+require "pathname"
+
 require_relative "./version"
 require_relative "./dependency_helper"
-require_relative "./logging"
+module Langchain
+  class << self
+    attr_accessor :logger
+
+    attr_reader :root
+  end
+
+  @logger ||= ::Logger.new($stdout, level: :warn, formatter: ->(severity, datetime, progname, msg) { "[LangChain.rb] #{msg}\n" })
+
+  @root = Pathname.new(__dir__)
+
+  autoload :Loader, "langchain/loader"
+
+  module Processors
+    autoload :Base, "langchain/processors/base"
+    autoload :CSV, "langchain/processors/csv"
+    autoload :Docx, "langchain/processors/docx"
+    autoload :HTML, "langchain/processors/html"
+    autoload :JSON, "langchain/processors/json"
+    autoload :JSONL, "langchain/processors/jsonl"
+    autoload :PDF, "langchain/processors/pdf"
+    autoload :Text, "langchain/processors/text"
+  end
+end
 
 module Agent
   autoload :Base, "agent/base"
@@ -11,8 +37,10 @@ end
 
 module Vectorsearch
   autoload :Base, "vectorsearch/base"
+  autoload :Chroma, "vectorsearch/chroma"
   autoload :Milvus, "vectorsearch/milvus"
   autoload :Pinecone, "vectorsearch/pinecone"
+  autoload :Pgvector, "vectorsearch/pgvector"
   autoload :Qdrant, "vectorsearch/qdrant"
   autoload :Weaviate, "vectorsearch/weaviate"
 end
@@ -20,8 +48,10 @@ end
 module LLM
   autoload :Base, "llm/base"
   autoload :Cohere, "llm/cohere"
+  autoload :GooglePalm, "llm/google_palm"
   autoload :HuggingFace, "llm/hugging_face"
   autoload :OpenAI, "llm/openai"
+  autoload :Replicate, "llm/replicate"
 end
 
 module Prompt