+++ lib/llm/openai.rb
@@ -18,9 +18,12 @@ module LLM
       @client = ::OpenAI::Client.new(access_token: api_key)
     end
 
+    #
     # Generate an embedding for a given text
+    #
     # @param text [String] The text to generate an embedding for
     # @return [Array] The embedding
+    #
     def embed(text:)
       response = client.embeddings(
         parameters: {
@@ -31,9 +34,12 @@ module LLM
       response.dig("data").first.dig("embedding")
     end
 
+    #
     # Generate a completion for a given prompt
+    #
     # @param prompt [String] The prompt to generate a completion for
     # @return [String] The completion
+    #
     def complete(prompt:, **params)
       default_params = {
         model: DEFAULTS[:completion_model_name],
@@ -51,15 +57,18 @@ module LLM
       response.dig("choices", 0, "text")
     end
 
+    #
     # Generate a chat completion for a given prompt
+    #
     # @param prompt [String] The prompt to generate a chat completion for
     # @return [String] The chat completion
+    #
     def chat(prompt:, **params)
       default_params = {
         model: DEFAULTS[:chat_completion_model_name],
         temperature: DEFAULTS[:temperature],
         # TODO: Figure out how to introduce persisted conversations
-        messages: [{ role: "user", content: prompt }]
+        messages: [{role: "user", content: prompt}]
       }
 
       if params[:stop_sequences]
@@ -72,6 +81,24 @@ module LLM
       response.dig("choices", 0, "message", "content")
     end
 
-    alias_method :generate_embedding, :embed
+    #
+    # Generate a summary for a given text
+    #
+    # @param text [String] The text to generate a summary for
+    # @return [String] The summary
+    #
+    def summarize(text:)
+      prompt_template = Prompt.load_from_path(
+        file_path: Langchain.root.join("llm/prompts/summarize_template.json")
+      )
+      prompt = prompt_template.format(text: text)
+
+      complete(
+        prompt: prompt,
+        temperature: DEFAULTS[:temperature],
+        # Most models have a context length of 2048 tokens (except for the newest models, which support 4096).
+        max_tokens: 2048
+      )
+    end
   end
 end