+++ spec/vectorsearch/base_spec.rb
@@ -30,6 +30,16 @@ RSpec.describe Vectorsearch::Base do
       ).to be_a(LLM::HuggingFace)
     end
 
+    it "correctly with llm: :replicate" do
+      expect(
+        described_class.new(
+          llm: :replicate,
+          llm_api_key: "123"
+        )
+        .llm_client
+      ).to be_a(LLM::Replicate)
+    end
+
     it "throws an error with currently unsupported llm: :anthropic" do
       expect {
         described_class.new(
@@ -78,6 +88,42 @@ RSpec.describe Vectorsearch::Base do
         ---
         Answer:
       PROMPT
+    end
+  end
+
+  describe "#add_data" do
+    it "allows adding single path" do
+      expect(subject).to receive(:add_texts).with(texts: array_with_strings_matcher(size: 1)) # not sure I love doing this
+
+      subject.add_data(path: Langchain.root.join("../spec/fixtures/loaders/cairo-unicode.pdf"))
+    end
+
+    it "allows adding multiple paths" do
+      paths = [
+        Langchain.root.join("../spec/fixtures/loaders/cairo-unicode.pdf"),
+        Langchain.root.join("../spec/fixtures/loaders/clearscan-with-image-removed.pdf"),
+        Langchain.root.join("../spec/fixtures/loaders/example.txt")
+      ]
+
+      expect(subject).to receive(:add_texts).with(texts: array_with_strings_matcher(size: 3)) # not sure I love doing this
+
+      subject.add_data(paths: paths)
+    end
+
+    it "requires path or paths" do
+      expect { subject.add_data }.to raise_error(ArgumentError, /Either path or paths must be provided/)
+    end
+
+    it "requires only path or paths" do
+      expect { subject.add_data(path: [], paths: []) }.to raise_error(ArgumentError, /Either path or paths must be provided, not both/)
+    end
+
+    def array_with_strings_matcher(size:)
+      proc do |array|
+        array.is_a?(Array) &&
+          array.length == size &&
+          array.all? { |e| e.is_a?(String) }
+      end
     end
   end
 end